backend/requirements.txt
psycopg2-binary==2.9.9
Faker==20.1.0
Django==4.2.*              
djangorestframework==3.15.* # API REST
python-dotenv==1.0.*        # carrega .env
openai==1.30.*              # IA que gera o resumo
drf-yasg==1.21.*          # Documentação da API